refactor(RandomNumberGenerator.kt): RandomNumberGenerator override NumberGenerator
